move python scripts into a subdir, fix paths in compose integration files
