+++ docker/docker-compose.integration-mysql.yml
@@ -19,4 +19,4 @@ services:
     ports:
       - "3307:3306"
     volumes:
-      - ./data/sql/mysql_example.sql:/docker-entrypoint-initdb.d/mysql_example.sql
+      - ../data/sql/mysql_example.sql:/docker-entrypoint-initdb.d/mysql_example.sql
