+++ scripts/generate_openapi.py
